Fixed big custom characters
I noticed that there was a problem drawing the "C" in "JUICE" with the original version. This is because only up to eight custom characters can be assigned to the  HD44780, any more definitions then "wrap around", so that custom character "8" points to custom character "0", custom character "9" points to custom character "1" and so on... So, there are two characters too many. The solution:

A blank (8) is already defined by space, or use character 254;
The block (7) can be replaced by other custom characters, to give more stylised lettering, or use character 255.
The cross custom character moved from code 9 to code 7.